% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closest_val_to_Pprob.R
\name{closest_val_to_Pprob}
\alias{closest_val_to_Pprob}
\title{Translate a value to the closest percentile probability in a vector}
\usage{
closest_val_to_Pprob(x, value)
}
\arguments{
\item{x}{vector with values}

\item{value}{cut off value for which the closest percentile in vector x needs
to be determined}
}
\value{
probability \link{0,1} which can be used in polarPlot (multiply by 100)
}
\description{
The polarPlot function from openair requires a probability between 0 and 100.
When multiple sites are evaluated it is best to keep the cut-off value the
same. However, contributions for sites are different, therefore the
percentiles also differ. To keep this value the same, this function can be
used to find the corresponding probability in other vectors.
}
