% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_obs_mod.R
\name{compare_obs_mod}
\alias{compare_obs_mod}
\title{Compare the observed total mass vs the modelled contributions in the G-matrix}
\usage{
compare_obs_mod(
  data,
  x = "pm10",
  y = "sum_factors",
  group = NA,
  robust = FALSE,
  maxit = 200,
  intercept = FALSE,
  mod.line = FALSE,
  regress.details = TRUE,
  CI = FALSE,
  CI.type = "area",
  CI.level = 0.95,
  CI.color = "dodgerblue4",
  CI.alpha = 0.4,
  CI.linetype = "dotted",
  CI.linewidth = 1,
  PI = FALSE,
  PI.type = "area",
  PI.level = 0.95,
  PI.color = "dodgerblue4",
  PI.linetype = "dashed",
  PI.linewidth = 1,
  PI.alpha = 0.4,
  show.plot = TRUE,
  line.color = "dodgerblue4",
  line.size = 1,
  point.color = "red",
  point.shape = 16,
  point.size = 2,
  point.alpha = 0.8,
  ylab = "Modelled",
  xlab = "Observed",
  auto.text = TRUE,
  asp.ratio.1 = FALSE,
  n.col = 2,
  geomtextpath = tibble::tibble(labels = list(`0.5` = "0.5", `1.0` = "1.0", `2.0` =
    "2.0"), hjust = list(`0.5` = 0.95, `1.0` = 0.95, `2.0` = 0.95))
)
}
\arguments{
\item{data}{Tibble containing at least two columns of data: 1) the
measured mass concentrations (i.e., PM2.5 or PM10) and 2) the sum of the
factors in concentration units. Additional information can be present, but
will not be used.}

\item{x}{The name of the column in the "data" tibble containing the
observations for the regression.}

\item{y}{The name of the column in the "data" tibble containing the
sum of all factors for the regression.}

\item{group}{Grouping variable, used to split data in subsets and create
facetted plots. Works well with regression per season or per year. Only
one grouping variable can be used.}

\item{robust}{Boolean used to switch between normal (OLS) or robust (RLS)
regression.}

\item{maxit}{The limit on the number of IWLS iterations used when robust
regression is performed. Defaults to 200. Can be increased in case of error
messages related to failed convergence in the number of steps used.}

\item{intercept}{Should an intercept be used when doing the regression?
Default is FALSE, forcing the regression through the origin.}

\item{mod.line}{This option will add three lines to assist with model
evaluations. These three lines are the 1:1 line (solid) and the 1:0.5 and
1:2 lines (dashed). Adding these lines will help to evaluate how close the
points are to the 1:1 relation. Points within the dashed lines will be
within a factor of two of the regression line.}

\item{regress.details}{Show the equation and R2 (non-robust only) on the
LR plot. Default usage is \code{TRUE} or \code{FALSE}.}

\item{CI}{Should the confidence levels (CI) be shown? Default of this
parameter is set to \dQuote{FALSE}.}

\item{CI.type}{How should the CI be displayed? There are two options:
\dQuote{area} and \dQuote{line}. With \dQuote{area} a shaded area is
created, whereas \dQuote{line} just plots the upper and lower limits of the
interval.}

\item{CI.level}{= Confidence level, between [0,1], defaults to
\dQuote{0.95} corresponding to 95\%.}

\item{CI.color}{The color of the CI area or line, depending on the
\code{CI.type}. Defaults to "dodgerblue4".}

\item{CI.alpha}{The alpha transparency for the CI area. Is ignored when
\code{PI.type = "area"}}

\item{CI.linetype}{The type of the lines. Is ignored when
\code{CI.type = "area"}}

\item{CI.linewidth}{The width of the lines. Is ignored when
\code{CI.type = "area"}}

\item{PI}{= Should the prediction levels (PI) be shown? Default of this
parameter is set to \dQuote{FALSE}.}

\item{PI.type}{How should the PI be displayed? See \code{CI.type} for more
info.}

\item{PI.level}{= Confidence level, between [0,1], defaults to
\dQuote{0.95} corresponding to 95\%.}

\item{PI.color}{= The color of the PI area or line, depending on the
\code{PI.type}. Defaults to "dodgerblue4".}

\item{PI.linetype}{The type of the lines. Is ignored when
\code{PI.type = "area"}}

\item{PI.linewidth}{The width of the lines. Is ignored when
\code{PI.type = "area"}}

\item{PI.alpha}{=  The alpha transparency for the PI area. Is ignored when
\code{PI.type = "area"}}

\item{show.plot}{A logical argument evaluating to TRUE or FALSE indicating
whether the plot should be shown as default.}

\item{line.color}{The color of the regression line. Defaults to
"dodgerblue4".}

\item{line.size}{The thickness of the regression line. Defaults to 1.}

\item{point.color}{The color of the points. This can be either a color (i.e.,
"red") or a column in the data. Defaults to "red".}

\item{point.shape}{The shape of the points. Defaults to 16.}

\item{point.size}{The size of the points. Defaults to 2.}

\item{point.alpha}{The alpha transparency for the points. Defaults to 0.25.}

\item{ylab}{y-axis label.}

\item{xlab}{x-axis label.}

\item{auto.text}{Either \code{TRUE} (default) or \code{FALSE}. If
\code{TRUE} attempts will be made to automatically format titles,
axis labels, pollutant names and units properly, e.g., by subscripting
the \sQuote{2} in NO2.}

\item{asp.ratio.1}{Output the plot with an aspect ratio of 1. For this to
work the minimum and maximum of x and y axis are used.}

\item{n.col}{The number of columns used in the facetted plot.}

\item{geomtextpath}{This variable contains specific settings for the optional
package geomtextpath. When plotting the mod.lines, the geomtextpath
package will add labels to the lines for clarification, but only if this
package is installed. If it is not installed, the function will only plot
the lines. The variable contains the labels and the hjust for positioning
the labels. The hjust shifts the label up and down the line, and is
entered as a value between 0-1. The labels are character labels that are
plotted as is.}
}
\value{
me2tools list containing the ggplot2 for the regression of the
observations and the sum of the factors in concentration units (LR). The
results of the LR models are also given, as well as the data used to
calculate these models. Additional information is provided regarding the
use of robust regression.
}
\description{
Based on the observed total mass and the modelled contributions in the
G-matrix this function performs a simpe linear regression on the observations
and the sum of the factors. The function can perform an ordinary least
squares regression (OLS) and a robust regression. In the latter regression
outliers have less influence on the determination of the regression line.
Also the regression can be based on a model with and without intercept. For
usage of multilinear regression (MLR) see function
\code{\link{compare_obs_mod_mlr}}
}
\seealso{
\code{\link{compare_obs_mod_mlr}}
}
