% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_obs_mod_mlr.R
\name{compare_obs_mod_mlr}
\alias{compare_obs_mod_mlr}
\title{Compare the observed total mass vs the modelled contributions in the G-matrix}
\usage{
compare_obs_mod_mlr(
  measurements,
  obs = "pm10",
  G_matrix,
  mod = c("factor_01", "factor_02"),
  factor_mass = NA,
  robust = FALSE,
  maxit = 200,
  intercept = FALSE,
  mod.line = FALSE,
  regress.details = TRUE,
  CI = TRUE,
  CI.type = "area",
  CI.level = 0.95,
  CI.color = "dodgerblue4",
  CI.alpha = 0.4,
  CI.linetype = "dotted",
  CI.linewidth = 1,
  PI = TRUE,
  PI.type = "area",
  PI.level = 0.95,
  PI.color = "dodgerblue4",
  PI.linetype = "dashed",
  PI.linewidth = 1,
  PI.alpha = 0.4,
  show.plot = TRUE,
  line.color = "dodgerblue4",
  line.size = 1,
  point.color = "red",
  point.shape = 16,
  point.size = 2,
  point.alpha = 0.8,
  ylab = "Modelled",
  xlab = "Observed",
  auto.text = TRUE,
  asp.ratio.1 = FALSE,
  n.col = 2
)
}
\arguments{
\item{measurements}{Tibble containing the date and the observations against
which a regression should be performed. Typical usage would be to provide
a tibble containing the date and the total mass (i.e., PM2.5 or PM10).}

\item{obs}{The name of the column in the "measurement" tibble containing the
observations for the regression.}

\item{G_matrix}{The preliminary output from functions reading G values in a
non-tidied format. In order to perform MLR this matrix needs to have
"normalized" factor contributions. If this matrix also contains the factor
contributions in "concentration" units, then these results are used to
calculate the sum of factors. This sum will then be used in the final
regression agains the measured mass. Note that this will note use the output
from the MLR nor the \code{factor_mass} to calculate factor contributions
in concentration units based on these results (i.e., provided factor
contributions always take precedence).}

\item{mod}{Character vector containing the names of the columns in the G
matrix which are considered the factors and should be used in the regression.}

\item{factor_mass}{This is a vector with the same length as the number of
factor profiles. It contains the total mass in concentration units which is
used to transform the G matrix from unity to concentration units. Providing
this factor will take precedence over the MLR results. If the MLR results
should be used, then this factor should be set to \code{NA}, which is also
the default}

\item{robust}{Boolean used to switch between normal (OLS) or robust (RLS)
regression for both the MLR as the LR.}

\item{maxit}{The limit on the number of IWLS iterations used when robust
regression is performed. Defaults to 200. Can be increased in case of error
messages related to failed convergence in the number of steps used.}

\item{intercept}{Should an intercept be used when doing the LR regression
(i.e., the sum of factors vs measured mass)? Default is FALSE, forcing
the regression through the origin. Note that the MLR is forced through the
origin by default.}

\item{mod.line}{This option will add three lines to assist with model
evaluations. These three lines are the 1:1 line (solid) and the 1:0.5 and
1:2 lines (dashed). Adding these lines will help to evaluate how close the
points are to the 1:1 relation. Points within the dashed lines will be
within a factor of two of the regression line.}

\item{regress.details}{Show the equation and R2 (non-robust only) on the
LR plot. Default usage is \code{TRUE} or \code{FALSE}.}

\item{CI}{Should the confidence levels (CI) be shown? Default of this
parameter is set to \dQuote{FALSE}.}

\item{CI.type}{How should the CI be displayed? There are two options:
\dQuote{area} and \dQuote{line}. With \dQuote{area} a shaded area is
created, whereas \dQuote{line} just plots the upper and lower limits of the
interval.}

\item{CI.level}{= Confidence level, between [0,1], defaults to
\dQuote{0.95} corresponding to 95\%.}

\item{CI.color}{The color of the CI area or line, depending on the
\code{CI.type}. Defaults to "dodgerblue4".}

\item{CI.alpha}{The alpha transparency for the CI area. Is ignored when
\code{PI.type = "area"}}

\item{CI.linetype}{The type of the lines. Is ignored when
\code{CI.type = "area"}}

\item{CI.linewidth}{The width of the lines. Is ignored when
\code{CI.type = "area"}}

\item{PI}{= Should the prediction levels (PI) be shown? Default of this
parameter is set to \dQuote{FALSE}.}

\item{PI.type}{How should the PI be displayed? See \code{CI.type} for more
info.}

\item{PI.level}{= Confidence level, between [0,1], defaults to
\dQuote{0.95} corresponding to 95\%.}

\item{PI.color}{= The color of the PI area or line, depending on the
\code{PI.type}. Defaults to "dodgerblue4".}

\item{PI.linetype}{The type of the lines. Is ignored when
\code{PI.type = "area"}}

\item{PI.linewidth}{The width of the lines. Is ignored when
\code{PI.type = "area"}}

\item{PI.alpha}{=  The alpha transparency for the PI area. Is ignored when
\code{PI.type = "area"}}

\item{show.plot}{A logical argument evaluating to TRUE or FALSE indicating
whether the plot should be shown as default.}

\item{line.color}{The color of the regression line. Defaults to
"dodgerblue4".}

\item{line.size}{The thickness of the regression line. Defaults to 1.}

\item{point.color}{The color of the points. Defaults to "red".}

\item{point.shape}{The shape of the points. Defaults to 16.}

\item{point.size}{The size of the points. Defaults to 2.}

\item{point.alpha}{The alpha transparency for the points. Defaults to 0.25.}

\item{ylab}{y-axis label.}

\item{xlab}{x-axis label.}

\item{auto.text}{Either \code{TRUE} (default) or \code{FALSE}. If
\code{TRUE} attempts will be made to automatically format titles,
axis labels, pollutant names and units properly, e.g., by subscripting
the \sQuote{2} in NO2.}

\item{asp.ratio.1}{Output the plot with an aspect ratio of 1. For this to
work the minimum and maximum of x and y axis are used.}

\item{n.col}{The number of columns used in the facetted plot.}
}
\value{
me2tools list containing the ggplot2 for the regression of the
observations and the sum of the factors in concentration units (LR). The
results of both the MLR and the LR models are also given, as well as the
data used to calculate these models. Additional information is provided
regarding the use of the concentration units in G and if robust regression
has been applied.

@section Using factor_mass for the sum of factors regression:
To perform the regression against the sum of factors, the factors in the
G-matrix also have to be available in the concentration units. By default,
only the "normalised" data is available, and a \code{factor_mass} is used to
transform this data into concentration units.

If the provided G-matrix only contains normalised data, a vector of the same
length of the number of factors can be provided in this function to
calculate the concentration units.

If this vector, called \code{factor_mass}, is not provided and no
concentration units are present in G, then the coefficients of the MLR are
used as the \code{factor_mass}.

If concentration units are present in G, then this will take precedence,
unless the \code{factor_mass} is defined. In the latter case, the existing
concentration data is overwritten with the new concentrations based on the
provided \code{factor_mass}.

In the output of the function, a variable is used to denote the application
of \code{factor_mass} and the possible concentrations in G matrix in the
regression.
}
\description{
Based on the observed total mass and the modelled contributions in the
G-matrix this function performs a multilinear regression (MLR) on the
observations and the individual normalised factor contributions. From this
regression the coefficient of each factor shows the contribution of these
factors to the total mass. If the variable \code{factor_mass} is not
provided, then the results of the MLR is used to transform the normalized
factor contributions towards concentration contributions. The sum of the
contributions of factors in concentration units, either calculated in the
function or provided in the function call, are then used in a regression
against the measured mass. For usage of multilinear regression (MLR) see
function \code{\link{compare_obs_mod_mlr}}
}
\seealso{
\code{\link{compare_obs_mod}}
}
