% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epa_percentile.R
\name{epa_percentile}
\alias{epa_percentile}
\title{Percentiles using weighted average approach}
\usage{
epa_percentile(x, prob, na.rm = TRUE)
}
\arguments{
\item{x}{vector for which the percentile needs to be calculated.}

\item{prob}{probability, ranges from 0 - 1 (i.e. 0.95 calculates the P95)}

\item{na.rm}{remove NA values}
}
\value{
requested percentile
}
\description{
\code{epa_percentile} calculates the percentiles using the weighted
average approach as outlined in the EPA PMF manual.
}
