% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epa_plot_BSboxplot.R
\name{epa_plot_BSboxplot}
\alias{epa_plot_BSboxplot}
\title{Plot BS Boxplot using the EPA-PMF style}
\usage{
epa_plot_BSboxplot(
  BS_results,
  xlabel.angle = 45,
  xlabel.order = NA,
  y_min = -5,
  ylab = "Concentration of species",
  xlab = "Species",
  auto.text = TRUE,
  x.font.size = 10,
  box.color = "burlywood1",
  outlier.color = "firebrick",
  outlier.size = 2,
  outlier.shape = 3,
  base.color = "royalblue",
  base.size = 2,
  base.shape = 23,
  show.plot = TRUE,
  rm.grid.x = FALSE,
  facet.parse.label = FALSE,
  ...
)
}
\arguments{
\item{BS_results}{Tibble containing the F-matrix results for the BS runs.
A list, containing data, as provided by \code{me2_BS_read_F} is also accepted.}

\item{xlabel.angle}{What angle should the x-axis labels be presented in? If
your labels are long, \code{45} degrees can be useful, which is also the
default.}

\item{xlabel.order}{The labels containing the species on the x-axis are
plotted based on the factor levels. This parameter can contain the levels
and the species are transformed to factors using these levels. N.B. these
levels can contain code for expressions.}

\item{y_min}{The minimum value for the y-axis, which is used to cut-off the
logarithmic scale. The default setting is \code{-5}, which corresponds to
1xE-5.}

\item{ylab}{Primary y-axis label.}

\item{xlab}{x-axis label.}

\item{auto.text}{Either \code{TRUE} (default) or \code{FALSE}. If
\code{TRUE} attempts will be made to automatically format titles,
axis labels, pollutant names and units properly, e.g., by subscripting
the \sQuote{2} in NO2.}

\item{x.font.size}{The size of the xtick labels. Defaults to 10.}

\item{box.color}{Provide the fill color of the IQR box. Is set to
\sQuote{burlywood1}, as an approximation of the color used in EPA-PMF
program.}

\item{outlier.color}{The fill color for the outliers, with the
default being \sQuote{firebrick}, similar to the color used in the EPA-PMF
program.}

\item{outlier.size}{The size of the point of the outliers,
defaults to 2.}

\item{outlier.shape}{The shape of the point of the outliers. The
default is 3 (+).}

\item{base.color}{The fill color for the base results, with the
default being \sQuote{royalblue}, similar to the color used in the EPA-PMF
program.}

\item{base.size}{The size of the point for the base results,
defaults to 2.}

\item{base.shape}{The shape of the point for the base results. The
default is 23 (diamond).}

\item{show.plot}{A logical argument evaluating to TRUE or FALSE indicating
whether the plot should be shown as default.}

\item{rm.grid.x}{Should the vertical grid lines be removed? In some cases
there are a lot of species, causing the vertical grid lines to clutter the
plot. If set to \code{TRUE} these lines are removed. Defaults to
\code{FALSE}.}

\item{facet.parse.label}{Should the labels be parsed using the
\code{labeller = label_parsed}? If set to \code{TRUE} then \code{"SO[2]"}
will use subscript on the labels shown in the facet.}
}
\value{
me2tools list containing the ggplot2 with box plot results for the
selected BS runs and the call to produce the plots.
}
\description{
This function plots a box and whisker plot based on the information read for
all BS runs using \code{me2_BS_read_F}. Note, that unlike the aggregated
results calculated by \code{me2_BS_read_F}, no filtering is applied by
default by this function. This means that each factor in each BS result is
used in the creation of the boxplot, even if their correlation is below
the defined correlation threshold. Factors with a correlation below the
threshold need to be removed manually based on the output provided by
\code{me2_BS_read_F}
}
