% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epa_plot_contribution.R
\name{epa_plot_contribution}
\alias{epa_plot_contribution}
\title{Plot contributions using the EPA-PMF style}
\usage{
epa_plot_contribution(
  G_matrix,
  unit = "concentration",
  xlabel.angle = 45,
  xlabel.order = NA,
  ylab = "Concentration of species",
  xlab = "Date",
  auto.text = TRUE,
  x.font.size = 10,
  show.plot = TRUE,
  rm.grid.x = FALSE,
  facet.parse.label = FALSE,
  ...
)
}
\arguments{
\item{G_matrix}{Tibble containing the G-matrix results for a single base run
in normalised or concentration units.}

\item{unit}{Which units should be plotted? By default the units of the
G_matrix are \dQuote{normalised} to 1. The other option, which is also the
default, is \dQuote{concentration}}

\item{xlabel.angle}{What angle should the x-axis labels be presented in? If
your labels are long, \code{45} degrees can be useful, which is also the
default.}

\item{xlabel.order}{The labels containing the species on the x-axis are
plotted based on the factor levels. This parameter can contain the levels
and the species are transformed to factors using these levels. N.B. these
levels can contain code for expressions.}

\item{ylab}{Primary y-axis label.}

\item{xlab}{x-axis label.}

\item{auto.text}{Either \code{TRUE} (default) or \code{FALSE}. If
\code{TRUE} attempts will be made to automatically format titles,
axis labels, pollutant names and units properly, e.g., by subscripting
the \sQuote{2} in NO2.}

\item{x.font.size}{The size of the xtick labels. Defaults to 10.}

\item{show.plot}{A logical argument evaluating to TRUE or FALSE indicating
whether the plot should be shown as default.}

\item{rm.grid.x}{Should the vertical grid lines be removed? In some cases
there are a lot of species, causing the vertical grid lines to clutter the
plot. If set to \code{TRUE} these lines are removed. Defaults to
\code{FALSE}.}

\item{facet.parse.label}{Should the labels be parsed using the
\code{labeller = label_parsed}? If set to \code{TRUE} then \code{"SO[2]"}
will use subscript on the labels shown in the facet.}

\item{...}{Other parameters, for example renamed parameters.}
}
\value{
me2tools list containing the ggplot2 with contributions
plot and the call used to create this plot.
}
\description{
Function to plot the G_matrix contributions for a single run into faceted
plots.
}
