% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epa_plot_QQexp.R
\name{epa_plot_QQexp}
\alias{epa_plot_QQexp}
\title{Plot Q/Qexp plot using the EPA-PMF style}
\usage{
epa_plot_QQexp(
  residuals,
  num_factors = NA,
  num_strong_species = NA,
  threshold = 2,
  xlabel.angle = 45,
  xlabel.vjust = 1,
  xlabel.hjust = 1,
  xlabel.order = NA,
  ylab = "Q/Qexp",
  auto.text = TRUE,
  date.breaks = NA,
  x.font.size = 10,
  x.n.dodge = 1,
  bar.color = "cadetblue3",
  bar.width = 0.9,
  bar.alpha = 0.8,
  point.color = "red",
  point.size = 1,
  point.alpha = 1,
  threshold.color = "black",
  threshold.size = 1,
  threshold.type = 2,
  show.plot = TRUE,
  show.legend = FALSE,
  rm.grid.x = FALSE,
  ...
)
}
\arguments{
\item{residuals}{Tibble from \code{\link[=me2_read_residuals]{me2_read_residuals()}} containing
the results for the scaled residuals. In order to use this function only
the results from a single base run should be provided.}

\item{num_factors}{Compulsory number of factors used to generate the
residuals.}

\item{num_strong_species}{Number of species that are designated as "strong"
variables in the analysis. Defaults to \code{NA}, meaning all species are
set to "strong".}

\item{threshold}{The threshold that should be used to check the Q/Qexp
against. Defaults to 2, as mentioned in the EPA-PMF manual as being the
threshold to observe species or samples that have not been well modelled.}

\item{xlabel.angle}{What angle should the x-axis labels be presented in? If
your labels are long, \code{45} degrees can be useful, which is also the
default.}

\item{xlabel.vjust}{vertical justification of the xlabel, between [0,1].
The default is NA, so that ggplot uses some heuristics to pick the best
value for this parameter. Any other value is processed by changing theme
settings,}

\item{xlabel.hjust}{horizontal justification of the xlabel, between [0,1].
The default is NA, so that ggplot uses some heuristics to pick the best
value for this parameter. Any other value is processed by changing theme
settings,}

\item{xlabel.order}{The labels containing the species on the x-axis are
plotted based on the factor levels. This parameter can contain the levels
and the species are transformed to factors using these levels. N.B. these
levels can contain code for expressions.}

\item{ylab}{Primary y-axis label.}

\item{auto.text}{Either \code{TRUE} (default) or \code{FALSE}. If
\code{TRUE} attempts will be made to automatically format titles,
axis labels, pollutant names and units properly, e.g., by subscripting
the \sQuote{2} in NO2.}

\item{date.breaks}{When this parameter is a string it is passed to the
function \code{scales::breaks_width()} and deals with the number of x
ticks. It contains the distance between each break. Either a number, or for
date/times, a single string of the form "{n} {unit}", e.g. "1 month",
"5 days". Unit can be of one "sec", "min", "hour", "day", "week",
"month", "year". It can also be a single integer, in which case the
variable is passed to \code{scales::breaks_pretty()}.}

\item{x.font.size}{The size of the xtick labels. Defaults to 10.}

\item{x.n.dodge}{The number of rows on the x-axis that should be used to
render the labels. Useful for labels that otherwise overlap, for example
with a large number of species. Also allows for a larger
\sQuote{x.font.size}.}

\item{bar.color}{Provide the fill color for the Q/Qexp based bars for each
species. Is set to \sQuote{cadetblue3}, as an approximation
of the color used in EPA-PMF program.}

\item{bar.width}{The width of the bar, expressed as a value between [0,1].}

\item{bar.alpha}{The alpha transparency of the fill color of the bar,
expressed as a value between [0,1].}

\item{point.color}{The color for the Q/Qexp point of the samples, with the
default being \sQuote{red}.}

\item{point.size}{The size for the Q/Qexp point of the samples,
defaults to 1.}

\item{point.alpha}{The alpha for the Q/Qexp point of the samples,
defaults to 1.}

\item{threshold.color}{The color used for plotting the threshold as a
horizontal line.}

\item{threshold.size}{The line size of the horizontal threshold line.
Defaults to 1.}

\item{threshold.type}{The line type of the horizontal threshold line.
Defaults to 2 (dashed line).}

\item{show.plot}{A logical argument evaluating to TRUE or FALSE indicating
whether the plot should be shown by default.}

\item{show.legend}{A logical argument evaluating to TRUE or FALSE indicating
whether the legend should be shown by default.}

\item{rm.grid.x}{Should the vertical grid lines be removed? In some cases
there are a lot of species, causing the vertical grid lines to clutter the
plot. If set to \code{TRUE} these lines are removed. Defaults to
\code{FALSE}.}

\item{...}{Other parameters, for example renamed parameters.}
}
\value{
me2tools list containing the ggplot2 with segment and point
geometries and the call.
}
\description{
Using this function the residuals of the PMF solution are plotted in a
Q/Qexp plot (both profiles and contributions). This is a great way to
examine the residuals, specifically the species or samples that were not well
modelled (i.e., Q/Qexp > 2).
}
