% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epa_plot_residuals.R
\name{epa_plot_residuals}
\alias{epa_plot_residuals}
\title{Function to plot histograms of the residuals with limit values.}
\usage{
epa_plot_residuals(
  me2_residuals,
  limit.species = NA,
  residual.limits = c(-4, 4),
  xlab = "Scaled residuals",
  ylab = "Number of residuals in group",
  auto.text = TRUE,
  bin.number = 30,
  bin.color = "steelblue",
  bin.alpha = 0.7,
  guide.color = "red",
  guide.linetype = "solid",
  guide.linewidth = 1,
  x.font.size = 10,
  xlabel.angle = 45,
  facet.col = 3,
  show.plot = TRUE,
  rm.grid.x = FALSE,
  facet.parse.label = FALSE
)
}
\arguments{
\item{me2_residuals}{The output of the function \code{me2_read_residuals},
which reads the data file with residual results}

\item{limit.species}{A character vector containing the names of species that
should be plotted. This is ideal to create bigger plots for examining the
species. Defaults to \code{NA} so no limit is applied.}

\item{residual.limits}{The limits of the scaled residuals. These limits are
plotted using vertical lines in the histogram.}

\item{xlab}{x-axis label.}

\item{ylab}{y-axis label.}

\item{auto.text}{Either \code{TRUE} (default) or \code{FALSE}. If
\code{TRUE} attempts will be made to automatically format titles,
axis labels, pollutant names and units properly, e.g., by subscripting
the \sQuote{2} in NO2.}

\item{bin.number}{The number of bins used, defaults to 30,}

\item{bin.color}{The color of the bins, defaults to "steelblue". This is
also the fill color for the box and whisker and violin plots.}

\item{bin.alpha}{The transparency of the bins, defaults to 0.7. This is
also the alpha for the box and whisker and violin plots.}

\item{guide.color}{= The color of the guidelines based on the
residual.limits. Defaults to "red" for all the plots.}

\item{guide.linetype}{= The line type of the guidelines based on the
residual.limits. Defaults to "solid"  for all the plots.}

\item{guide.linewidth}{The line width of the guidelines based on the
residual.limits. Defaults to 1.}

\item{x.font.size}{The size of the xtick labels. Defaults to 10.}

\item{xlabel.angle}{What angle should the x-axis labels be presented in? If
your labels are long, \code{45} degrees can be useful, which is also the
default.}

\item{facet.col}{The number of columns for the faceted plot. If the number of
columns is set to 1, the faceting will be in rows only. Defaults to 3 and
is only applied to the histogram plot.}

\item{show.plot}{A logical argument evaluating to TRUE or FALSE indicating
whether the histogram plot should be shown as default.}

\item{rm.grid.x}{Should the vertical grid lines be removed? In some cases
there are a lot of species, causing the vertical grid lines to clutter the
plot. If set to \code{TRUE} these lines are removed. Defaults to
\code{FALSE}.}

\item{facet.parse.label}{Should the labels be parsed using the
\code{labeller = label_parsed}? If set to \code{TRUE} then \code{"SO[2]"}
will use subscript on the labels shown in the facet.}
}
\value{
me2tools list containing the ggplot2 with scaled residual results
for the selected species in a histogram, a box and whisker plot of the
same data, a violin plot of the same data, the residual limits and the
call to produce the plot.
}
\description{
This function plots the results of the scaled residuals using histograms.
}
