% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epa_sn.R
\name{epa_sn}
\alias{epa_sn}
\title{S/N ratio according to the EPA-PMF V5 manual}
\usage{
epa_sn(x, x_unc, na.rm = TRUE)
}
\arguments{
\item{x}{vector containing the measurements}

\item{x_unc}{vector containing the measurement uncertainty associated with
the measurements}

\item{na.rm}{remove NA values}
}
\value{
S/N ratio
}
\description{
\code{epa_sn} calculates the S/N ratio using the approach as outlined in
the EPA PMF manual.
}
