% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_overview.R
\name{input_overview}
\alias{input_overview}
\title{Overview of the input of ME2}
\usage{
input_overview(
  me2_input,
  limit.species = NA,
  sn.colors = list(bad = "firebrick3", weak = "goldenrod2", strong = "green4"),
  sn.limits = list(u_limit = 1, l_limit = 0.5),
  xlab = "Concentration (ug/m3)",
  ylab = "Uncertainty (ug/m3)",
  auto.text = TRUE,
  x.font.size = 10,
  xlabel.angle = 45,
  facet.col = 3,
  show.plot = TRUE,
  rm.grid.x = FALSE,
  facet.parse.label = FALSE
)
}
\arguments{
\item{me2_input}{The output of the function \code{me2_read_input}, which
reads an input file defined with a date-time column, followed by
alternating concentrations and uncertainties for a number of unknown
species}

\item{limit.species}{A character vector containing the names of species that
should be plotted. This is ideal to create bigger plots for examining the
species. Defaults to \code{NA} so no limit is applied.}

\item{sn.colors}{This parameter is a list containing the colors used in the
concentration vs uncertainty plots. In these plots the calculated S/N ratio
is used to color the points}

\item{sn.limits}{These are the limits for the S/N ratio. Values above the
\code{u_limit} are classified as "strong", the values below the
\code{l_limit} are classified as "bad" and the values between the limits
are classified as "weak". The default limits are \code{u_limit = 1} and
\code{l_limit = 0.5} which correspond to the S/N calculation introduced in
EPA-PMF version 5. Note this is only a classification and does not mean
that data should be excluded or included by default.}

\item{xlab}{x-axis label.}

\item{ylab}{y-axis label.}

\item{auto.text}{Either \code{TRUE} (default) or \code{FALSE}. If
\code{TRUE} attempts will be made to automatically format titles,
axis labels, pollutant names and units properly, e.g., by subscripting
the \sQuote{2} in NO2.}

\item{x.font.size}{The size of the xtick labels. Defaults to 10.}

\item{xlabel.angle}{What angle should the x-axis labels be presented in? If
your labels are long, \code{45} degrees can be useful, which is also the
default.}

\item{facet.col}{The number of columns for the faceted plot. If the number of
columns is set to 1, the faceting will be in rows only. Defaults to 3.}

\item{show.plot}{A logical argument evaluating to TRUE or FALSE indicating
whether the plot should be shown as default.}

\item{rm.grid.x}{Should the vertical grid lines be removed? In some cases
there are a lot of species, causing the vertical grid lines to clutter the
plot. If set to \code{TRUE} these lines are removed. Defaults to
\code{FALSE}.}

\item{facet.parse.label}{Should the labels be parsed using the
\code{labeller = label_parsed}? If set to \code{TRUE} then \code{"SO[2]"}
will use subscript on the labels shown in the facet.}
}
\value{
me2tools list containing the ggplot2 with concentration vs
uncertainty for the selected species and the call to produce the plot.
}
\description{
This function uses the input for ME2 to provide an overview of the species
concentrations and uncertainties.
}
