% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_color.R
\name{is_color}
\alias{is_color}
\title{Check if the string is an actual color representation}
\usage{
is_color(x)
}
\arguments{
\item{x}{a string or character vector to test if they represent strings}
}
\value{
a named logical vector containing TRUE or FALSE
}
\description{
This function checks if the input, either a string (i.e. "red") or a
character vector contains strings that represent a color. Code is based on
information from https://stackoverflow.com/questions/13289009/check-if-character-string-is-a-valid-color-representation
}
