% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_factor_order.R
\name{match_factor_order}
\alias{match_factor_order}
\title{Match factors between base and source based on Pearson correlation}
\usage{
match_factor_order(
  base,
  source,
  F.profile = "percentage_of_species_sum",
  G.unit = "normalised",
  corr.threshold = 0.6,
  extended.output = FALSE
)
}
\arguments{
\item{base}{base F.matrix or base G.matrix. This is only one matrix from
a single run.}

\item{source}{source F.matrix or source G.matrix. This function can only
handle comparison between one other matrix from a single run.}

\item{F.profile}{what version of the F matrix should be used for the
correlations, i.e., concentration units or percentage of species sum.
Defaults to \code{percentage_of_species_sum}}

\item{G.unit}{what version of the G matrix should the correlations be
performed on? Normalized (avg = 1) or concentration units? Defaults to
\code{normalised}}

\item{corr.threshold}{value for the Pearson correlation threshold, defaults
to 0.6 as stipulated by the EPA-PMF 5.0 manual.}

\item{extended.output}{provide extended output (besides the factor output
the correlation matrix is also provided when this value is set to
\code{TRUE})}
}
\value{
factor output (named vector). If \code{extended.output} is
\code{TRUE} a list is returned with the factor output and the correlation
matrix.

Note, in order to use this function to assess the mapping of multiple sources
for F or G matrix, the user should use this function inside a loop.

The function can also be used to assess the correlations of F and G factors
with partly overlapping species (F) or samples (G). If different amount of
species or samples are found (i.e., non-overlapping species/samples), these
are added to the matrices that do not have these and their contribution is
set to 0 by default. If the user wishes to compare only those species or
samples present in both base and source, the input should be modified as
such prior to running this function.
}
\description{
With this function the Pearson correlation coefficient between the source
and base F or G matrix can be calculated. It can be used to assign a
bootstrap (BS) run factor to a base factor and observe the number of
unmapped factors. Or it can be used to map the results of any other run
towards the selected base case with the lowest Q or Qmain contribution.
}
