% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/me2_BSDISP_read_res.R
\name{me2_BSDISP_read_res}
\alias{me2_BSDISP_read_res}
\title{Get the F values from the BSDISP results file (BSDISPres?.txt)}
\usage{
me2_BSDISP_read_res(
  BSDISPres_file,
  base_run = 1,
  tidy_output = FALSE,
  species = NA,
  dc_species = NA
)
}
\arguments{
\item{BSDISPres_file}{location and file name of the \dQuote{DISPres_?.txt} file,
corresponding to either dQmax = [0.5, 1, 2, 4].}

\item{base_run}{The number of the base run associated with the BSDISP results}

\item{tidy_output}{Should the output be reshaped into tidy data? Default:
FALSE}

\item{species}{A vector containing the names of the species for the rows in
the F-matrix. If these species name are outputted in the ME-2 output as the
second column (a column of row numbers being the first), then these values
are used when \code{species = NA}. If this second column with names is not
available all species are named as \dQuote{species_xx}, with xx being an
unique number starting at 1.}

\item{dc_species}{character vector containing the species that should be
subtracted from the missing mass to account for double counting. See also
the section about double counting. This will only be applied to the
concentration values}
}
\value{
\code{me2_BSDISP_read_res} returns an object of class ``me2tools''.
  The object includes three main components: \code{call}, the command used
  to read the data; \code{data}, the BSDISP data for each BS run;
  and \code{F_format}, the aggregated BSDISP data in the same format as the
  F_matrix. If retained, e.g., using 
  \code{output <- me2_BSDISP_read_res(file)}, this output can be used to
  recover the data, reproduce, or undertake further analysis.

  An me2tools output can be manipulated using a number of generic operations,
  including \code{print}, \code{plot} and \code{summary}.
}
\description{
Files stored after a BSDISP run are named with a user-specific prefix, shown
here as an asterisk (*). Three output files (*_BSDISP.dat,  *_BSDISP.txt and
*_BSDISP.rsd) are stored after the DISP phase of the BSDISP run (note that
the BS phase of the BS-DISP has outputs with a \dQuote{_BS_} suffix, i.e.,
*_BS_.dat, *_BS_.txt, and *_BS_.rsd) and currently there are no functions
available for reading information from the \dQuote{txt} file. The residuals,
stored in the *_BSDISP.rsd file, can be read using \code{me2_read_residuals}.
Besides these three files, four other files are produced as output,
corresponding to dQmax = [0.5, 1, 2, 4] and are: *_BSDISPres1.txt,
*_BSDISPres2.txt, *_BSDISPres3.txt and *_DISPres4.txt, which can be read with
this function. These files contain the results for each BS resample after
displacing both up and down. In this function the results are also
aggregated to \dQuote{BSDISP_P05}, \dQuote{BSDISP_P95} and
\dQuote{BSDISP_avg}, based on the BSDISP results in concentration units.
}
\section{Adding species to F}{

By far the easiest way to add \dQuote{species} to F is to provide them as
input parameters. The species names can probably be found in the original
data input used for ME-2 calculations.

Other ways to replace the \dQuote{species} might be by using conditional
replacement of the default names or the use of solutions involving
\code{rep} to replicate elements of vectors.
}

\section{Renaming factor profiles}{

As the labels of the factor profiles are unknown, this routine reads them as
\dQuote{factor_xx}, with xx being a unique number and outputted as a
character array. The easiest way to work with this data is by re-coding this
vector as \code{factor}. Then the order of the factor profiles and the
correct names can be easily set using the following code.

\preformatted{
mydata$factor <- factor(mydata$factor)
mydata$factor <- dplyr::recode_factor(mydata$factor,
                                      `factor_01` = "MyFirstName",
                                      `factor_02` = "MySecondName",
                                      ...
)
}
}

\section{Double counting in missing mass variable}{


In some ME-2 applications a variable containing the missing mass is used.
This variable often contains the difference between the total mass (i.e.
PM10) and the sum of constituents measured in the total mass. Typically,
in factor analysis it is assumed that the sum of constituents is
approximately equal to the total mass. In other words, the constituents are
responsible for most, if not all, of the total mass.

If this is not the case, there are non-measured constituents that contribute
to the total mass. As such, the missing mass (denoted as \dQuote{m.mass}
in me2tools) is calculated by subtracting the sum of the constituents from
the total mass and introduced in the factor analysis.

However, in the case of multi-time (MT) factor analysis it might be
difficult to estimate the \dQuote{m.mass} due to the various time resolutions
in the data. Often only part of the available constituents can be summed and
subtracted from total mass (if they have the same sampling interval/time
resolution), leaving an overestimation of the \dQuote{m.mass}. After factor
analysis, this overestimation in the \dQuote{m.mass} still exists. Summing
the results of the factor analysis for each constituents, including
\dQuote{m.mass}, will therefore also be an overestimation due to double
counting of some masses. As such, comparison against the total mass will
fail.

To prevent this overestimation the \dQuote{m.mass} variable should be
corrected for double counting. A viable way to do this in MT-factor analysis
is to subtract the sum of constituents that should have been used in the
initial calculation of \dQuote{m.mass} but could not be applied due to
different time resolutions. Since the F matrix from the factor analysis
is provided with unified time resolutions, the mass associated with these
constituents in the F matrix can now be subtracted to provide a better
estimate of \dQuote{m.mass}.

The variable \dQuote{dc_species} is a character vector containing
constituents that should be subtracted from the \dQuote{m.mass} variable to
account for double counting.
}

\seealso{
\code{\link{me2_read_F}}, \code{\link{me2_BS_read_F}},
\code{\link{me2_DISP_read_F}}, \code{\link{me2_DISP_read_res}},
\code{\link{me2_read_all}}, \code{\link{me2_read_dat}}
}
