% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_suggests.R
\name{me2_check_suggests}
\alias{me2_check_suggests}
\title{Check for missing suggested packages}
\usage{
me2_check_suggests()
}
\value{
Invisibly returns a character vector of missing package names.
}
\description{
This function checks whether the specified packages (typically suggested ones)
are installed on the user's system. If any are missing, it displays a helpful
message recommending how to install them.
}
\examples{
\dontrun{
me2_check_suggests()
}

}
