% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/me2_DISP_read_res.R
\name{me2_DISP_read_minmax}
\alias{me2_DISP_read_minmax}
\title{Get the minimum and maximum DISP values from the DISP results file
(DISPres_x.txt)}
\usage{
me2_DISP_read_minmax(DISPres_file, tidy_output = FALSE, species = NA)
}
\description{
This function has been renamed to \code{me2_DISP_read_res}.
}
\seealso{
\code{\link{me2_DISP_read_res}}
}
