% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/me2_DISP_read_res.R
\name{me2_DISP_read_res}
\alias{me2_DISP_read_res}
\title{Get the minimum and maximum DISP values from the DISP results file
  (DISPres_x.txt)}
\usage{
me2_DISP_read_res(
  DISPres_file,
  base_run = 1,
  tidy_output = FALSE,
  species = NA
)
}
\arguments{
\item{DISPres_file}{location and file name of the \dQuote{DISPres_?.txt} file,
corresponding to either dQmax = [4, 8, 16, 32].}

\item{base_run}{The number of the base run associated with the DISP results}

\item{tidy_output}{Should the output be reshaped into tidy data? Default:
FALSE}

\item{species}{A vector containing the names of the species for the rows in
the F-matrix. If these species names are outputted in the ME-2 output as 
the second column (a column of row numbers being the first), then these 
values are used when \code{species = NA}. If this second column with names 
is not available, all species are named as \dQuote{species_xx}, with xx 
being an unique number starting at 1.}
}
\value{
\code{me2_DISP_read_res} returns an object of class ``me2tools''.
  The object includes five main components: \code{call}, the command used
  to read the data; \code{data}, the DISP data for each BS run;
  \code{F_format}, the aggregated DISP data in the same format as the
  F_matrix; \code{diag} the diagnostic line from the file, containing the error code (0
  is no error, 6 or 9 indicates that the run was aborted) and the largest
  drop of Q; and \code{swaps}, the swap counts for each dQmax level (4, 8, 
  15, 25). If retained, e.g., using 
  \code{output <- me2_BSDISP_read_res(file)}, this output can be used to
  recover the data, reproduce, or undertake further analysis.

  An me2tools output can be manipulated using a number of generic operations,
  including \code{print}, \code{plot} and \code{summary}.
}
\description{
Files stored after a DISP run are named with a user-specific prefix, shown 
here as an asterisk (*). Three output files (*_DISP.dat,  *_DISP.txt and 
*_DISP.rsd) are stored after a DISP run. The function \code{me2_DISP_read_F} 
reads all the factor profiles in the text file and \code{me2_DISP_read_G} 
reads all the factor contributions. The residuals, stored in the 
*_DISP.rsd file, can be read using \code{ me2_read_residuals}. Besides these 
three files, four other files are produced as output, corresponding to 
dQmax = [4, 8, 16, 32] and are *_DISPres1.txt, *_DISPres2.txt, 
*_DISPres3.txt and *_DISPres4.txt. These files contain the minimum and 
maximum DISP results for a specific dQmax and any of these files can be read 
using this function. Each of these files have the same format, containing 
minimum and maximum DISP values in concentration units and as percentage of 
species units. These are all read and stored in the output.
}
\section{Adding species to min/max}{

By far the easiest way to add \dQuote{species} to the DISP min/max results is
to provide them as input parameters. The species names can probably be found
in the original data input used for ME-2 calculations.

Other ways to replace the \dQuote{species} might be by using conditional
replacement of the default names or the use of solutions involving
\code{rep} to replicate elements of vectors.
}

\section{Renaming factor profiles}{

As the labels of the factor profiles are unknown, this routine reads them as
\dQuote{factor_xx}, with xx being a unique number and outputted as a
character array. The easiest way to work with this data is by re-coding this
vector as \code{factor}. Then the order of the factor profiles and the
correct names can be easily set using the following code.

\preformatted{
mydata$factor <- factor(mydata$factor)
mydata$factor <- dplyr::recode_factor(mydata$factor,
                                      `factor_01` = "MyFirstName",
                                      `factor_02` = "MySecondName",
                                      ...
)
}

Please note that the above will only work when the data is read with the
\code{tidy_output = TRUE} setting.
}

\seealso{
\code{\link{me2_read_F}}, \code{\link{me2_BS_read_F}}, 
\code{\link{me2_DISP_read_F}}, \code{\link{me2_read_all}}, 
\code{\link{me2_read_dat}}
}
