% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/me2_modify_ini.R
\name{me2_modify_ini}
\alias{me2_modify_ini}
\title{Modify ME-2 ini file using R}
\usage{
me2_modify_ini(
  file,
  openfile = NA,
  robust = NA,
  n2 = NA,
  ns = NA,
  np = NA,
  numtasks = NA,
  fileid = NA,
  aasmooc1 = NA,
  aasmooc3 = NA,
  c2 = NA,
  c3 = NA,
  em = NA
)
}
\arguments{
\item{file}{name of the ME-2 ini file, can include path. This is also the
file where the output is written to.}

\item{openfile}{name of the data file for ME-2}

\item{robust}{use robust calculations}

\item{n2}{number of variables}

\item{ns}{number of samples}

\item{np}{number of factors}

\item{numtasks}{number of tasks}

\item{fileid}{part of the file name of the output files generated by ME-2}

\item{aasmooc1}{value for c1 smoothing}

\item{aasmooc3}{value for c3 smoothing}

\item{c2}{value for c2}

\item{c3}{value for c3}

\item{em}{the error model used in ME-2

In the example a "template.ini" is used. This is basically an ini file that
contains most of the parameters for the run we want to do. We can use the
function to modify the ini, save it as a new file and use that as input
for ME-2.}
}
\value{
overwriting the provided \code{ini} file with the updated version.
}
\description{
The .ini file, which is needed to  provide information to ME-2, can be
programmatically changed using this function. Not all information can be
changed yet.
}
\examples{
\dontrun{
factors <- c(5,6,7,8)

for (factor in factors) {
   me2_modify_ini(file = "template.ini",
                  robust = 1,
                  np = factor,
                  n2 = 16 #species,
                  ns = 1024 #samples,
                  aasmooc1 = 2.5,
                  aasmooc3 = 0.15,
                  c3 = 0.15,
                  numtasks = 250,
                  openfile = "my_input_data.txt",
                  fileid = paste0("my_output_file_", factor))
   # run model
   shell("me2 template.ini")
   # manually rename LOG so it fits with the rest of the files
   file.rename("ME2.LOG", paste0("my_output_file_", factor, ".LOG"))
}
}

}
