% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/me2_read_input.R
\name{me2_read_input}
\alias{me2_read_input}
\title{Get the input data}
\usage{
me2_read_input(
  file,
  unc_identifier = c(".std", "_std", "_unc", ".unc", "U_", "U.", "unc_", "unc."),
  sep = "\\t",
  tz = "Etc/GMT-1"
)
}
\arguments{
\item{file}{File name of the multi time data file.}

\item{unc_identifier}{Vector containing the prefix/suffix to denote the
uncertainty for a species. Examples are \code{xx_std}, \code{xx_unc},
\code{U_xx} and so on. The function splits the data into a data and
uncertainty set by splitting the columns that contain one or more of the
\code{unc_identifier}. Default is
\code{c(".std", "_std", "_unc", ".unc", "U_", "U.", "unc_", "unc.")}.}

\item{sep}{The column separator used to separate values on one line. The
default is \code{"\t"} which is a TAB character. Other options are
\code{";"} or \code{","}}

\item{tz}{Time zone for date time, defaults to \code{Etc/GMT-1}}
}
\value{
list containing full data, both concentrations and uncertainties.
}
\description{
The data from the input data file can be read using this function. It
also performs some clean up, allowing the use of data in R.
}
\section{Data format}{

The format of the multi time data has to consists of the following columns,
in the same order as given in the table.

\tabular{ll}{
\strong{Name}    \tab \strong{Description}\cr
DateTime:        \tab Start date of the measurement in the 'yyyy-mm-dd hh:mm(:ss)' or 'dd-mm-yyyy hh:mm(:ss)' format\cr
...              \tab Columns of species with their appropriate names, alternating concentrations and uncertainties for each specie (i.e. specie1, unc_specie1, specie2, ...)\cr
}
}

