% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/me2_read_log.R
\name{me2_read_log}
\alias{me2_read_log}
\title{Get data from the ME-2 log file}
\usage{
me2_read_log(
  me2_log_file,
  block_boundaries = list(start =
    "^\\\\s*Starting to process with the same .INI file next task:")
)
}
\arguments{
\item{me2_log_file}{file name of the ME-2 log file.}

\item{block_boundaries}{A list containing the \dQuote{start} string used
to identify the boundaries of the block containing the values
for the log file.}
}
\value{
tibble containing the following information over each base_run:
\dQuote{convergence} (T/F); \dQuote{self.cancel} (T/F);
\dQuote{iter.steps} (#); \dQuote{precon_mode} (#) and \dQuote{messages}.
}
\description{
The Multilinear Engine, version 2 (ME-2), has several options for default
logging. This function tries to read the messages from the log file. Please
note that inside the .ini file you can set the amount of output for the log
file. In some cases these outputs are very verbose and cannot be used in this
function. In this function the messages inside the LOG file are read using the
following header information, denoting the \dQuote{start} and \dQuote{end}
lines of the block containing the required data:
}
\seealso{
\code{\link{me2_read_all}}
}
