% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/me2_read_Q.R
\name{me2_read_Q}
\alias{me2_read_Q}
\title{Get the Q values from the results file (.txt)}
\usage{
me2_read_Q(
  me2_txt_file,
  block_boundaries = list(start = "^\\\\s*Sum-of-squares  Q, Qmain, Qaux")
)
}
\arguments{
\item{me2_txt_file}{ME2 output file (.txt), containing the results and
auxiliary information.}

\item{block_boundaries}{A list containing the \dQuote{start} string used to
identify the boundaries of the block containing the values for the
residuals. For residuals the \dQuote{end} string is automatically derived
based on the \dQuote{start} string for the next block.}
}
\value{
tibble containing the Q values for each run.
}
\description{
Typical use of Multilinear Engine version 2 (ME-2) will provide three
different files after a run: \dQuote{.dat}: machine readable format,
\dQuote{.rsd}: results for the residuals and \dQuote{.txt}: text file with
auxiliary information (i.e. headers). The headers in the text (.txt) file
are used in me2tools to split the data into several blocks. In this function
the Q values are read using user provided header information, denoting the
\dQuote{start} line of the block containing the required data.
}
\seealso{
\code{\link{me2_read_all}}
}
