% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/me2_read_Qcontrib.R
\name{me2_read_Qcontrib}
\alias{me2_read_Qcontrib}
\title{Get the values contributing to the Qmain from the results file (.txt)}
\usage{
me2_read_Qcontrib(
  me2_txt_file,
  block_boundaries = list(start = "^\\\\s*Contribution in Qmain by different variables",
    end = "^\\\\s*Task #,  seed:")
)
}
\arguments{
\item{me2_txt_file}{ME2 output file (.txt), containing the results and
auxiliary information.}

\item{block_boundaries}{A list containing the \dQuote{start} and
\dQuote{end} string used to identify the boundaries of the block
containing the Qmain values.}
}
\value{
tibble containing Qmain and contributions to Qmain for multiple runs
}
\description{
Typical use of Multilinear Engine version 2 (ME-2) will provide three
different files after a run: \dQuote{.dat}: machine readable format,
\dQuote{.rsd}: results for the residuals and \dQuote{.txt}: text file with
auxiliary information (i.e. headers). The headers in the text (.txt) file
are used in me2tools to split the data into several blocks. In this function
the values contributing to Qmain are read using user provided header
information, denoting the \dQuote{start} and \dQuote{end} lines of the block
containing the required data.
}
\seealso{
\code{\link{me2_read_all}}
}
