% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/me2_read_residuals.R
\name{me2_read_residuals}
\alias{me2_read_residuals}
\title{Get the residuals from the residual file}
\usage{
me2_read_residuals(
  me2_res_file,
  tidy_output = FALSE,
  block_boundaries = list(start = "^\\\\s*Scaled Residuals"),
  species = NA,
  dates = NA,
  tz = "Etc/GMT-1"
)
}
\arguments{
\item{me2_res_file}{ME2 output file (.rsd), containing the results for the
residuals.}

\item{tidy_output}{Should the output of both G and F be reshaped into tidy
data? Default: FALSE}

\item{block_boundaries}{A list containing the \dQuote{start} string used to
identify the boundaries of the block containing the values for the
residuals. For residuals the \dQuote{end} string is automatically derived
based on the \dQuote{start} string for the next block.}

\item{species}{A vector containing the names of the species for the columns
in the residuals.}

\item{dates}{A vector containing the sample dates for the rows in the
residuals.}

\item{tz}{Parameter to control the time zone when parameter \dQuote{dates} is
not used. Default: 'Etc/GMT-1'}
}
\value{
tibble containing scaled residuals for each run.
}
\description{
Typical use of Multilinear Engine version 2 (ME-2) will provide three
different files after a run: \dQuote{.dat}: machine readable format,
\dQuote{.rsd}: results for the residuals and \dQuote{.txt}: text file with
auxiliary information (i.e. headers). This function reads the residual file
and supports two different formats. The first format is based on the EPA_PMF
format which consists of a block of residuals and a block of scaled residuals
for each run. The second format is based on the multi-time script, which only
provides the scaled residuals. In this file the blocks are separated by a
header row. The contents of this header row can be provided through the
\code{block_boundaries} parameter. This parameter is ignored in cae of
reading the EPA-PMF formated results.
The column \code{residual_type} in the output denotes the type of residual,
being \dQuote{residual} and \dQuote{residual_scaled}.
}
\seealso{
\code{\link{me2_read_all}}
}
