% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/me2tools_check_packages.R
\name{me2tools_check_packages}
\alias{me2tools_check_packages}
\title{Check necessary packages for me2tools}
\usage{
me2tools_check_packages()
}
\value{
Returns boolean: \code{TRUE} if everything has been found and
\code{FALSE} when packages are missing.
}
\description{
Performs a check if all needed packages are actually installed on the system.
Will provide messages when missing packages are detected.
}
