% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metcor.ne.graticules}
\alias{metcor.ne.graticules}
\title{Natural Earth Graticules.}
\format{
An object of class \code{list} of length 3.
}
\source{
Natural Earth Data
}
\usage{
metcor.ne.graticules
}
\description{
This data set contains the grids at 1, 5, 10, 15, 20, and 30° intervals.
Includes WGS84 bounding box. Graticules include 180-degree east and west
meridians (consider deleting one of these to avoid duplicate lines).
}
\details{
Data is stored in a list for the "small", "medium" and "large" scale.

Data extracted from Natural Earth Data (https://www.naturalearthdata.com/downloads/50m-physical-vectors/50m-graticules/)
}
\keyword{datasets}
