% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metcor_plot.R
\name{metcor_plot}
\alias{metcor_plot}
\title{Plot the MetCor results}
\usage{
metcor_plot(
  metcor.raster,
  metcor.plot.options = metcor_plot_options(),
  receptor = NA,
  verbose = TRUE,
  show.plot = TRUE
)
}
\arguments{
\item{metcor.raster}{MetCor raster object. Can be the default raster after
reading (see \link{metcor_import}) or a transformed projected raster (see
\link{metcor_project_raster})}

\item{metcor.plot.options}{List containing all the plot options (see
\link{metcor_plot_options})}

\item{receptor}{Tibble containing the location of the receptors with
longitude and latitude coordinates. Please note that the column names
should be "lat" and "lon".}

\item{verbose}{show detailed messages of what is happening. Defaults to
\code{TRUE}}

\item{show.plot}{A logical evaluating to TRUE or FALSE indicating whether the
plot should be outputed as default.}
}
\value{
\code{metcor_plot} returns an object of class ``me2tools''.
The object includes five main components: \code{call}, the command used
to generate the plot; \code{plot}, the default plot based on the settings;
\code{plot.options}, containing all information needed to generate the
plot, \code{proj_crs}, the used CRS and \code{origin}, the center point
used in the plot. If retained, e.g. using
\code{output <- metcor_plot(metcor.raster)}, this output can be used to
recover the data, reproduce or rework the original plot or undertake
further analysis.

An me2tools output can be manipulated using a number of generic operations,
including \code{print}, \code{plot} and \code{summary}.

@seealso \code{\link{metcor_import}}, \code{\link{metcor_project_raster}},
\code{\link{metcor_plot_options}}
}
\description{
This function plots the MetCor results. For the world map data from the
\code{rnaturalearth} package is used.
}
\seealso{
\code{\link{metcor_import}}, \code{\link{metcor_project_raster}},
\code{\link{metcor_plot_options}}, \code{\link{metcor_export}},
\code{\link{metcor_fix_hysplit}}
}
