% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metcor_project_raster.R
\name{metcor_project_raster}
\alias{metcor_project_raster}
\title{Project MetCor raster results}
\usage{
metcor_project_raster(
  raster,
  projection = "ESRI:102017",
  method = "bilinear",
  na.rm = TRUE
)
}
\arguments{
\item{raster}{Imported MetCor ASCII grid (see \link{metcor_import})}

\item{projection}{CRS for the projection (defaults to ESRI:102017)}

\item{method}{See \code{terra:project} for more details on method.}

\item{na.rm}{remove columns and rows with only NA values? Default is
\code{TRUE}}
}
\value{
projected raster object
}
\description{
This function changes the projection of the imported MetCor ASCII grid from
\code{EPSG:4326 (WGS 84)} to a new projection. Currently the only projection
supported is the \code{ESRI:102017} projection (North Pole Lambert Azimuthal
Equal Area)
}
\details{
The projected grid will always have the same resolution as the original grid.
}
\seealso{
\code{\link{metcor_import}}, \code{\link{metcor_plot}},
\code{\link{metcor_plot_options}}, \code{\link{metcor_export}},
\code{\link{metcor_fix_hysplit}}
}
