% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_contributions.R
\name{temporal_contributions}
\alias{temporal_contributions}
\title{Create temporal plots of contributions (G-matrix)}
\usage{
temporal_contributions(
  mydata,
  factor = NA,
  type = "default",
  group = NA,
  na.rm = TRUE,
  xlab = NA,
  ylab = NA,
  auto.text = TRUE,
  txt.x.rm = FALSE,
  txt.y.rm = FALSE,
  cols = "hue",
  whisk.lim = NULL,
  point.size = 1.25,
  line.size = 1,
  facet.col = 2,
  ...
)
}
\arguments{
\item{mydata}{A data frame of hourly (or higher temporal resolution data).
Must include a \code{date} field and at least one variable (factor
contributions) to plot.}

\item{factor}{Name of variable to plot, typically a factor from the G-matrix.}

\item{type}{\code{type} determines how the data are split i.e. conditioned,
and then plotted. The \dQuote{default} will produce a single plot using the
entire data. Type can be one of the built-in types as detailed in
\code{cutData} e.g. \dQuote{season}, \dQuote{year}, \dQuote{weekday} and
so on. For example, \code{type = "season"} will produce four plots --- one
for each season.

It is also possible to choose \code{type} as another variable in the data
frame. If that variable is numeric, then the data will be split into four
quantiles (if possible) and labelled accordingly. If type is an existing
character or factor variable, then those categories/levels will be used
directly. This offers great flexibility for understanding the variation of
different variables and how they depend on one another.}

\item{group}{This sets the grouping variable to be used. For example, if a
data frame had a column \code{site} setting \code{group = "site"} will
plot all sites together in each panel.}

\item{na.rm}{A logical evaluating to TRUE or FALSE indicating whether NA
values should be stripped before the computation proceeds.}

\item{xlab}{x-axis label.}

\item{ylab}{y-axis label.}

\item{auto.text}{Either \code{TRUE} (default) or \code{FALSE}. If
\code{TRUE} titles and axis labels will automatically try and format
pollutant names and units properly e.g.  by subscripting the \sQuote{2} in
NO2.}

\item{txt.x.rm}{A logical evaluating to TRUE or FALSE indicating whether the
x-axis tick labels should be removed.}

\item{txt.y.rm}{A logical evaluating to TRUE or FALSE indicating whether the
y-axis tick labels should be removed.}

\item{cols}{Colours to be used for plotting. Options include
\dQuote{default}, \dQuote{increment}, \dQuote{heat}, \dQuote{jet} and
\code{RColorBrewer} colours --- see the \code{openair} \code{openColours}
function for more details. For user defined the user can supply a list of
colour names recognised by R (type \code{colours()} to see the full list).
An example would be \code{cols = c("yellow", "green", "blue")}}

\item{whisk.lim}{User defined limits for the whiskers. Default is
\code{NULL}, in this case the whiskers calculated by \dQuote{geom_boxplot}
are used. In other cases a vector should be provided with two values
between 0..1 denoting the percentile (i.e., \code{c(0.05, 0.95)} will use
the P05 and P95 for the whisker limits).}

\item{point.size}{The size of the points. Defaults to 1.25.}

\item{line.size}{The size of the lines. Defaults to 1.}

\item{facet.col}{The number of columns for the faceted plot. Defaults to 2.}

\item{...}{Other parameters passed onto \code{cutData}. For example, in the
case of \code{cutData} the option \code{hemisphere = "southern"}.}
}
\value{
\code{temporal_factors} returns an object of class ``me2tools''.
The object includes three main components: \code{call}, the command used
to generate the plot; \code{plot}, the default plot based on the settings;
and \code{box.plot}, the associated box and whisker plots in the style of
Tukey.  If retained, e.g. using \code{output <- temporal_factors(mydata,
  "factor1")}, this output can be used to recover the data, reproduce or
rework the original plot or undertake further analysis.

An me2tools output can be manipulated using a number of generic operations,
including \code{print}, \code{plot} and \code{summary}.
}
\description{
This function can create temporal plots for the factors in the G-contribution
matrix. Default uses all statistics as calculated by geom_boxplot.
}
\seealso{
\code{\link[ggplot2]{geom_boxplot}}, \code{\link[openair]{cutData}}
}
