% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_obs_mod.R
\name{compare_obs_mod}
\alias{compare_obs_mod}
\title{Compare the observed total mass vs the modelled contributions in the G-matrix}
\usage{
compare_obs_mod(
  measurements,
  obs = "pm10",
  G_matrix,
  mod = c("factor_01", "factor_02"),
  factor_mass = NA,
  robust = FALSE,
  show.plot = TRUE,
  line.color = "dodgerblue4",
  line.size = 1,
  point.color = "red",
  point.shape = 16,
  point.size = 2,
  point.alpha = 0.25,
  ylab = "Modelled",
  xlab = "Observed",
  auto.text = TRUE
)
}
\arguments{
\item{measurements}{Tibble containing the date and the observations against
which a regression should be performed. Typical usage would be to provide
the total mass (i.e., PM2.5 or PM10).}

\item{obs}{The name of the column in the "measurement" tibble containing the
observations for the regression.}

\item{G_matrix}{The preliminary output from functions reading G values in a
non-tidied format.}

\item{mod}{Vector containing the names of the columns in the G matrix which
are considered the factors and should be used in the regression.}

\item{factor_mass}{This is a vector with the same length as the number of
factor profiles. It contains the total mass in concentration units which is
used to transform the G matrix from unity to concentration units.}

\item{robust}{Boolean used to switch between normal (OLS) or robust (RLS)
regression.}

\item{show.plot}{A logical argument evaluating to TRUE or FALSE indicating
whether the plot should be shown as default.}

\item{line.color}{The color of the regression line. Defaults to
"dodgerblue4".}

\item{line.size}{The thickness of the regression line. Defaults to 1.}

\item{point.color}{The color of the points. Defaults to "red".}

\item{point.shape}{The shape of the points. Defaults to 16.}

\item{point.size}{The size of the points. Defaults to 2.}

\item{point.alpha}{The alpha transparency for the points. Defaults to 0.25.}

\item{ylab}{y-axis label.}

\item{xlab}{x-axis label.}

\item{auto.text}{Either \code{TRUE} (default) or \code{FALSE}. If
\code{TRUE} attempts will be made to automatically format titles,
axis labels, pollutant names and units properly, e.g., by subscripting
the \sQuote{2} in NO2.}
}
\value{
me2tools list containing the ggplot2 for the regression of the
observations and the sum of the factors in concentration units (LR). The
results of both the MLR and the LR models are also given, as well as the
data used to calculate these models. Additionally information is provided
regarding the use of the concentration units in G and if robust regression
has been applied.

@section Using factor_mass for the sum of factors regression:
To perform the regression against the sum of factors, the factors in the
G-matrix also have to be available in the concentration units. By default,
only the "normalised" data is available, and a \code{factor_mass} is used to
transform this data into concentration units.

If the provided G-matrix only contains normalised data, a vector of the same
length of the number of factors can be provided in this function to
calculate the concentration units.

If this vector, called \code{factor_mass}, is not provided and no
concentration units are present in G, then the coefficients of the MLR are
used as the \code{factor_mass}.

If concentration units are present in G, than this will take precedence,
unless the \code{factor_mass} is defined. In the latter case the existing
concentration data is overwritten with the new concentrations based on the
provided \code{factor_mass}.

In the output of the function a variable is used to denote the application
of \code{factor_mass} and the possible concentrations in G matrix in the
regression.
}
\description{
Based on the observed total mass and the modelled contributions in the
G-matrix this function performs a multilinear regression on the observations
and the individual normalised factor contributions. The function also
performs a regression against the observations and the sum of the factors
}
