% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epa_plot_errorsummary.R
\name{epa_plot_errorsummary}
\alias{epa_plot_errorsummary}
\title{Plot error summary using the EPA-PMF style}
\usage{
epa_plot_errorsummary(
  F_matrix,
  xlabel.angle = 45,
  xlabel.order = NA,
  y_min = -5,
  ylab = "Concentration of species",
  xlab = "Species",
  auto.text = TRUE,
  x.font.size = 10,
  x.n.dodge = 1,
  bar.colors = list(BS = "goldenrod2", BSDISP = "green4", DISP = "royalblue2", base =
    "black"),
  bar.width = 0.9,
  show.plot = TRUE,
  show.legend = FALSE,
  expand.mult = c(0.015, 0.005),
  rm.grid.x = FALSE,
  facet.parse.label = FALSE,
  ...
)
}
\arguments{
\item{F_matrix}{Tibble containing the F-matrix results for the base runs
runs and at least one of the error estimates BS, DISP or BSDISP.}

\item{xlabel.angle}{What angle should the x-axis labels be presented in? If
your labels are long, \code{45} degrees can be useful, which is also the
default.}

\item{xlabel.order}{The labels containing the species on the x-axis are
plotted based on the factor levels. This parameter can contain the levels
and the species are transformed to factors using these levels. N.B. these
levels can contain code for expressions.}

\item{y_min}{The minimum value for the y-axis, which is used to cut-off the
logarithmic scale. The default setting is \code{-5}, which corresponds to
1xE-5.}

\item{ylab}{Primary y-axis label.}

\item{xlab}{x-axis label.}

\item{auto.text}{Either \code{TRUE} (default) or \code{FALSE}. If
\code{TRUE} attempts will be made to automatically format titles,
axis labels, pollutant names and units properly, e.g., by subscripting
the \sQuote{2} in NO2.}

\item{x.font.size}{The size of the xtick labels. Defaults to 10.}

\item{x.n.dodge}{The number of rows on the x-axis that should be used to
render the labels. Useful for labels that otherwise overlap, for example
with a large number of species. Also allows for a larger
\sQuote{x.font.size}.}

\item{bar.colors}{The bar colors for each error estimate (BS, DISP, BSDISP)
and the color for the average of the base run.}

\item{bar.width}{The width of the bar, expressed as a value between [0,1].}

\item{show.plot}{A logical argument evaluating to TRUE or FALSE indicating
whether the plot should be shown by default.}

\item{show.legend}{A logical argument evaluating to TRUE or FALSE indicating
whether the legend should be shown by default.}

\item{expand.mult}{Vector of multiplicative range expansion factors used on
the x-axis . Defaults to \code{c(0.015,0.005)}, as this seems to work best.
If length is 1, then both left and right x-axis are multiplied with this
factor. If length is 2, the left limit is expanded by \code{expand.mult[1]}
and the right by \code{expand.mult[2]}.}

\item{rm.grid.x}{Should the vertical grid lines be removed? In some cases
there are a lot of species, causing the vertical grid lines to clutter the
plot. If set to \code{TRUE} these lines are removed. Defaults to
\code{FALSE}.}

\item{facet.parse.label}{Should the labels be parsed using the
\code{labeller = label_parsed}? If set to \code{TRUE} then \code{"SO[2]"}
will use subscript on the labels shown in the facet.}

\item{...}{Other parameters, for example renamed parameters.}
}
\value{
me2tools list containing the ggplot2 with error estimates summary
plot and the call used to create this plot.
}
\description{
Function to plot the error summary using a combined F_matrix containing the
base run and at least one or all of the error estimates BS, DISP or BSDISP.
The results for the error estimates need to have a run_type of
\dQuote{DISP_min} and \dQuote{DISP_max} for DISP; the \dQuote{BS_P05} and
\dQuote{BS_P95} for BS and \dQuote{BSDISP_P05} and \dQuote{BSDISP_P95} for
BSDISP, respectively. For the average base_run concentrations the
concentration of each species from the base run is used.
}
