% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epa_plot_profile.R
\name{epa_plot_profile}
\alias{epa_plot_profile}
\title{Plot profiles using the EPA-PMF style}
\usage{
epa_plot_profile(
  F_matrix,
  xlabel.angle = 45,
  xlabel.vjust = 0.5,
  xlabel.hjust = 0.5,
  xlabel.order = NA,
  y_min = -5,
  ylab = "Concentration of species",
  ylab2 = "EV (\%)",
  xlab = "Species",
  auto.text = TRUE,
  x.font.size = 10,
  bar.color = "cadetblue3",
  bar.width = 0.9,
  point.color = "red",
  point.size = 3,
  lollipops = TRUE,
  disp.errorbar = FALSE,
  disp.errorbar.color = "orange",
  show.plot = TRUE,
  expand.mult = c(0.015, 0.005),
  rm.grid.x = FALSE
)
}
\arguments{
\item{F_matrix}{Tibble from \code{\link[pmfr]{me2_read_F}} containing
the results for the F_matrix. In order to use this function only the
results from a single run should be provided.}

\item{xlabel.angle}{What angle should the x-axis labels be presented in? If
your labels are long, \code{45} degrees can be useful, which is also the
default.}

\item{xlabel.vjust}{vertical justification of the xlabel, between [0,1].}

\item{xlabel.hjust}{horizontal justification of the xlabel, between [0,1].}

\item{xlabel.order}{The labels containing the species on the x-axis are
plotted based on the factor levels. This parameter can contain the levels
and the species are transformed to factors using these levels. N.B. these
levels can contain code for expressions.}

\item{y_min}{The minimum value for the y-axis, which is used to cut-off the
logarithmic scale. The default setting is \code{-5}, which corresponds to
1xE-5.}

\item{ylab}{Primary y-axis label.}

\item{ylab2}{Secondary y-axis label.}

\item{xlab}{x-axis label.}

\item{auto.text}{Either \code{TRUE} (default) or \code{FALSE}. If
\code{TRUE} attempts will be made to automatically format titles,
axis labels, pollutant names and units properly, e.g., by subscripting
the \sQuote{2} in NO2.}

\item{x.font.size}{The size of the xtick labels. Defaults to 10.}

\item{bar.color}{Provide the fill color for the concentration based bars on
the logarithmic scale. Is set to \sQuote{steelblue}, as an approximation of
the color used in EPA-PMF program.}

\item{bar.width}{The width of the bar, expressed as a value between 0-1.}

\item{point.color}{The fill color for the explained variation, with the
default being \sQuote{red}, similar to the color used in the EPA-PMF
program.}

\item{point.size}{The size of the point, typically the bar.width / 2}

\item{lollipops}{Should the dot representing the EV be extended with a line,
representing so it looks more like a lollipop? Defaults to \code{TRUE}.}

\item{disp.errorbar}{If the F_matrix contains the DISP min and max values,
the error bars can be plotted in this plot. Note that with this option the
species concentration is replaced by the DISP average.}

\item{disp.errorbar.color}{The color used for plotting the error bars for
DISP.}

\item{show.plot}{A logical argument evaluating to TRUE or FALSE indicating
whether the plot should be shown as default.}

\item{expand.mult}{Vector of multiplicative range expansion factors used on
the x-axis . Defaults to \code{c(0.015,0.005)}, as this seems to work best.
If length is 1, then both left and right x-axis are multiplied with this
factor. If length is 2, the left limit is expanded by \code{expand.mult[1]}
and the right by \code{expand.mult[2]}.}

\item{rm.grid.x}{Should the vertical grid lines be removed? In some cases
there are a lot of species, causing the vertical grid lines to clutter the
plot. If set to \code{TRUE} these lines are removed. Defaults to
\code{FALSE}.}
}
\value{
me2tools list containing the ggplot2 with segment and point
geometries and the call.
}
\description{
Function to plot PMF factor profiles using data from PMFR (using
\code{\link[pmfr]{read_pmf_factor_profiles}} and
\code{\link[pmfr]{tidy_pmf_profiles}}) or from the ME2tools package using
\code{\link{me2_read_F}} in the same way as in EPA-PMF.
}
