% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/me2_read_residuals.R
\name{me2_read_residuals}
\alias{me2_read_residuals}
\title{Get the residuals from the residual file}
\usage{
me2_read_residuals(
  me2_res_file,
  block_boundaries = list(start = "^\\\\s*Scaled Residuals"),
  species = NA,
  dates = NA,
  tz = "Etc/GMT-1"
)
}
\arguments{
\item{me2_res_file}{ME2 output file (.rsd), containing the results for the
residuals.}

\item{block_boundaries}{A list containing the \dQuote{start} string used to
identify the boundaries of the block containing the values for the
residuals. For residuals the \dQuote{end} string is automatically derived
based on the \dQuote{start} string for the next block.}

\item{species}{A vector containing the names of the species for the columns
in the residuals.}

\item{dates}{A vector containing the sample dates for the rows in the
residuals.}

\item{tz}{Parameter to control the time zone when parameter \dQuote{dates} is
not used. Default: 'Etc/GMT-1'}
}
\value{
tibble containing scaled residuals for each run.
}
\description{
Typical use of Multilinear Engine version 2 (ME-2) will provide three
different files after a run: \dQuote{.dat}: machine readable format,
\dQuote{.rsd}: results for the residuals and \dQuote{.txt}: text file with
auxiliary information (i.e. headers). The headers in file containing the
residual results (.rsd) file are used in me2tools to split the data into
several blocks. In this function the residuals are read using user provided
header information, denoting the \dQuote{start}line of the block containing
the required data.
}
\seealso{
\code{\link{me2_read_all}}
}
