% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metcor_export.R
\name{metcor_export}
\alias{metcor_export}
\title{Export G-matrix to MetCor}
\usage{
metcor_export(
  G_matrix,
  file,
  time_res = "auto",
  factor_names = "auto",
  force_utc = FALSE
)
}
\arguments{
\item{G_matrix}{G-matrix that should be exported to a MetCor input file. This
is typically the G_matrix provided by \code{\link{me2_read_G}} with the
\code{tidy_output} parameter set to \code{FALSE}. This output should be
expanded with the latitude (column: \code{lat}) and longitude (column:
\code{lon}) coordinates of the receptor site(s). These coordinates should
be exactly the same as the coordinates used in the \strong{HYSPLIT} model for
the receptor sites. Based on the unique lat and lon coordinates the
number of receptor sites is automatically derived.}

\item{file}{File name and location to save the MetCor input file}

\item{time_res}{The time resolution between the measurements (in minutes),
which is used to calculate the start (ITIME) and end times (FTIME) of each
sample needed in MetCor. These times are used to associate the correct
trajectories. If this variable is set to "auto" the function will try to
derive the time resolution automatically from the data. Note: the "auto"
feature only works when data sets of different time resolutions are not
mixed.}

\item{factor_names}{The names of the columns in the G-matrix that are
identified as factors. These columns will be used to add the
data/threshold values to output file. If \code{factor_names} is set to
"auto" the function tries to find the factors based on the known columns
in the G-matrix. Note, this feature will not work if non-default columns
are present in the G-matrix.}

\item{force_utc}{With this parameter the date time output for Metcor can be
forced into UTC (the same timezone as Hysplit results). When set to
\code{TRUE} there has to be no correction for the "Correlated Data Time
Zone" within Metcor. If this parameter is \code{FALSE}, there has to be
a manual correction when the time zone of the data is not UTC. Default
setting of this parameter is \code{FALSE}, so no correction is applied when
creating the output.}
}
\value{
tab-delimited plain text file containing date ranges, variable
names, data/threshold values, receptor site coordinates and an optional
multi site variable RECEPTORMAX if the number of receptor sites based on
the uniqueness of the lat/lon coordinates in the input is more than 1.

It should be noted that the full RECEPTORMAX, denoting the number of
receptor sites used in the analysis, is not required in MetCor. It is
perfectly fine to use a RECEPTORMAX smaller than the number of sites. In
this case, the user is advised to make manual changes to the output file.
The automatic default of RECEPTORMAX in MetCor is 2, when the RECEPTORMAX
is not given but the data is multi-site.
}
\description{
This function exports the G-matrix to a file that can be used as input for
MetCor.
}
\seealso{
\code{\link{metcor_import}}, \code{\link{metcor_project_raster}},
\code{\link{metcor_plot}}, \code{\link{metcor_plot_options}},
\code{\link{metcor_fix_hysplit}}
}
