% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metcor_import.R
\name{metcor_import}
\alias{metcor_import}
\title{Import MetCor results file into R}
\usage{
metcor_import(file, extent = "north", type = "wgs84", na.rm = TRUE)
}
\arguments{
\item{file}{MetCor results file}

\item{extent}{The extent of the grid. Valid values are \dQuote{north} for
the Northern hemisphere, \dQuote{south} for the Southern hemisphere and
\dQuote{worl} for the world.}

\item{type}{The type of grid, can be \dQuote{metcor} or \dQuote{WGS84}.
With the latter option the grid is transformed to a regular lat/lon grid
within the boundaries of [-180,180] for longitude and [-90,90] for
latitude. The \dQuote{metcor} option uses the MetCor transformed grid with
longitude set as [0,360].}

\item{na.rm}{remove columns and rows with only NA values? Default is
\code{TRUE}}
}
\value{
spatial object containing the results from MetCor.
}
\description{
This function reads the MetCor results file into a spatial raster object
(\code{CRS} = \code{EPSG:4326 (WGS 84)}).
Currently this is limited to the Northern Hemisphere (x=0:360, y=0:90)
}
\seealso{
\code{\link{metcor_project_raster}}, \code{\link{metcor_plot}},
\code{\link{metcor_plot_options}}, \code{\link{metcor_export}},
\code{\link{metcor_fix_hysplit}}
}
